Crate sp_transaction_pool

Transaction pool primitives types & Runtime API.

Modules

error

Transaction pool errors.

runtime_api

Tagged Transaction Queue Runtime API.

Structs

PoolStatus

Transaction pool status.

Enums

ChainEvent

Events that the transaction pool listens for.

TransactionSource

The source of the transaction.

TransactionStatus

Possible transaction status events.

Traits

InPoolTransaction

In-pool transaction interface.

LocalTransactionPool

Transaction pool interface for submitting local transactions that exposes a blocking interface for submission.

MaintainedTransactionPool

Trait for transaction pool maintenance.

OffchainSubmitTransaction

An abstraction for transaction pool.

TransactionPool

Transaction pool interface.

Type Definitions

BlockHash

Block hash type for a pool.

ImportNotificationStream

The import notification event stream.

LocalTransactionFor

Transaction type for a local pool.

PoolFuture

Typical future type used in transaction pool api.

TransactionFor

Transaction type for a pool.

TransactionLongevity

Minimum number of blocks a transaction will remain valid for. TransactionLongevity::max_value() means “forever”.

TransactionPriority

Priority for a transaction. Additive. Higher is better.

TransactionStatusStream

The stream of transaction events.

TransactionStatusStreamFor

Type of transactions event stream for a pool.

TransactionTag

Tag for a transaction. No two transactions with the same tag should be placed on-chain.

TxHash

Transaction hash type for a pool.