sp_transaction_pool::runtime_api

Trait TaggedTransactionQueue

pub trait TaggedTransactionQueue<Block: BlockT>: Core<Block> {
    // Provided methods
    fn validate_transaction_before_version_2(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        tx: <Block as BlockT>::Extrinsic,
    ) -> Result<TransactionValidity, ApiError> { ... }
    fn validate_transaction_before_version_3(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        source: TransactionSource,
        tx: <Block as BlockT>::Extrinsic,
    ) -> Result<TransactionValidity, ApiError> { ... }
    fn validate_transaction(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        source: TransactionSource,
        tx: <Block as BlockT>::Extrinsic,
        block_hash: Block::Hash,
    ) -> Result<TransactionValidity, ApiError> { ... }
}

The TaggedTransactionQueue api trait for interfering with the transaction queue.

Provided Methods

fn validate_transaction_before_version_2( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, tx: <Block as BlockT>::Extrinsic, ) -> Result<TransactionValidity, ApiError>

👎Deprecated

Validate the transaction.

fn validate_transaction_before_version_3( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, source: TransactionSource, tx: <Block as BlockT>::Extrinsic, ) -> Result<TransactionValidity, ApiError>

👎Deprecated

Validate the transaction.

fn validate_transaction( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, source: TransactionSource, tx: <Block as BlockT>::Extrinsic, block_hash: Block::Hash, ) -> Result<TransactionValidity, ApiError>

Validate the transaction.

This method is invoked by the transaction pool to learn details about given transaction. The implementation should make sure to verify the correctness of the transaction against current state. The given block_hash corresponds to the hash of the block that is used as current state.

Note that this call may be performed by the pool multiple times and transactions might be verified in any possible order.