Trait sp_transaction_pool::InPoolTransaction

pub trait InPoolTransaction {
    type Transaction;
    type Hash;
    fn data(&self) -> &Self::Transaction;

    fn hash(&self) -> &Self::Hash;

    fn priority(&self) -> &TransactionPriority;

    fn longevity(&self) -> &TransactionLongevity;

    fn requires(&self) -> &[TransactionTag];

    fn provides(&self) -> &[TransactionTag];

    fn is_propagable(&self) -> bool;
}

In-pool transaction interface.

The pool is container of transactions that are implementing this trait. See sp_runtime::ValidTransaction for details about every field.

Associated Types

type Transaction

Transaction type.

type Hash

Transaction hash type.

Loading content...
Required methods

fn data(&self) -> &Self::Transaction

Get the reference to the transaction data.

fn hash(&self) -> &Self::Hash

Get hash of the transaction.

fn priority(&self) -> &TransactionPriority

Get priority of the transaction.

fn longevity(&self) -> &TransactionLongevity

Get longevity of the transaction.

fn requires(&self) -> &[TransactionTag]

Get transaction dependencies.

fn provides(&self) -> &[TransactionTag]

Get tags that transaction provides.

fn is_propagable(&self) -> bool

Return a flag indicating if the transaction should be propagated to other peers.

Loading content...